requirements.txt
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
idna==2.8
Kivy==1.10.1
Kivy-Garden==0.1.4
numpy==1.16.2
opencv-contrib-python==4.0.1.24
opencv-python==4.0.0.21
Pillow==5.4.1
PyAudio==0.2.11
Pygments==2.4.0
pynput==1.4
pypiwin32==223
pywin32==224
requests==2.21.0
six==1.12.0
SpeechRecognition==3.8.1
tqdm==4.31.1
urllib3==1.24.3
